requirements.txt
streamlit>=1.30.0
yfinance>=0.2.30
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pywin32>=306
supabase>=2.4.0
python-dotenv>=1.0.0
streamlit-authenticator>=0.3.0
streamlit-autorefresh>=0.0.1
google-generativeai>=0.3.0
